+++ server/migrations/n4234567890-create-Catalogs.js
@@ -12,7 +12,7 @@ module.exports = {
         },
         userId: {
           allowNull: false,
-          type: Sequelize.SMALLINT,
+          type: Sequelize.INTEGER,
         },
         catalogName: {
           allowNull: false,
